Move createDueFilterOrErrorMessage() to bew class DueDateField
